Revert "refactor(device_detail): extracted shared widgets into their own files"
This reverts commit 1e6f11150d8d6977d3e4e5a6d6d84beb27bf57de.
